requirements.txt
Flask==3.0.1
gunicorn==23.0.0

requests==2.32.4
structlog==25.1.0

google-auth==2.3.2

camelot-py[cv]>=0.11.0
pandas>=2.0.0
opencv-python>=4.6.0
pypdf2>=3.0.0
